+++ src/debugger/breakpoints/breakpoints_source.h
@@ -85,7 +85,7 @@ class SourceBreakpoints
         std::vector<std::pair<std::string, bool>> logMessageParts;
         // In case of code line in constructor, we could resolve multiple methods for breakpoints.
         // For example, `MyType obj = new MyType(1);` code will be added to all class constructors).
-        std::vector<ToRelease<ICorDebugFunctionBreakpoint>> trFuncBreakpoints;
+        std::vector<std::pair<ToRelease<ICorDebugFunctionBreakpoint>, CORDB_ADDRESS>> trFuncBreakpoints;
 
         [[nodiscard]] bool IsVerified() const
         {